+++ packages/sharec/src/cli/spinner.js
@@ -5,9 +5,11 @@ const ora = require('ora')
  */
 
 /**
+ * Creates wrapper below ora spinner
+ * Can be silenced by parameters
  * @param {Object} params
- * @param {String} params.text
- * @param {Boolean} [params.silent]
+ * @param {String} params.text Spinner's initial text
+ * @param {Boolean} [params.silent] Ignores all spinner messages and frames
  * @returns {Spinner}
  */
 const createSpinner = ({ text, silent }) => {
@@ -20,6 +22,7 @@ const createSpinner = ({ text, silent }) => {
 
   return {
     /**
+     * Starts created spinner
      * @memberof Spinner
      * @returns {Spinner}
      */
@@ -34,6 +37,7 @@ const createSpinner = ({ text, silent }) => {
     },
 
     /**
+     * Checks spinner's current frame as succeed and returns spinner
      * @memberof Spinner
      * @param {String} text
      * @returns {Spinner}
@@ -49,6 +53,7 @@ const createSpinner = ({ text, silent }) => {
     },
 
     /**
+     * Checks spinner's current frame as failed and returns spinner
      * @memberof Spinner
      * @param {String} text
      * @returns {Spinner}
@@ -64,6 +69,7 @@ const createSpinner = ({ text, silent }) => {
     },
 
     /**
+     * Just changes spinner's current text
      * @memberof Spinner
      * @param {String} text
      * @returns {Spinner}