+++ packages/sharec/src/strategies/pipes/babel/test/pipe.test.js
@@ -5,6 +5,8 @@ describe('strategies > pipes > babel > pipe', () => {
   const babelBaseFxt = fixtures('babel/json/00-base')
 
   it('should merge babel json configs', () => {
-    expect(pipe('.babelrc')(babelBaseFxt)).toWraplessEqual(babelBaseFxt.result)
+    expect(pipe('.babelrc')(babelBaseFxt)).toWraplessEqual(babelBaseFxt.result, {
+      eol: false,
+    })
   })
 })