+++ packages/sharec/src/strategies/pipes/browserslist/test/pipe.test.js
@@ -5,6 +5,8 @@ describe('strategies > pipes > browserslist', () => {
   const browserslistBaseFxt = fixtures('browserslist/lines/00-base')
 
   it('should process browserslist configs', () => {
-    expect(pipe('.browserslistrc')(browserslistBaseFxt)).toWraplessEqual(browserslistBaseFxt.result)
+    expect(pipe('.browserslistrc')(browserslistBaseFxt)).toWraplessEqual(browserslistBaseFxt.result, {
+      eol: false,
+    })
   })
 })