+++ packages/sharec/src/test/install.test.js
@@ -10,14 +10,6 @@ describe('sharec > install', () => {
   const npmignoreFxt = fixtures('npmignore/lines/00-base')
   const gitignoreFxt = fixtures('gitignore/lines/00-base')
 
-  const targetProcess = {
-    argv: [null, null, 'install'],
-    env: {
-      INIT_CWD: '/target',
-    },
-    exit: jest.fn(),
-  }
-
   beforeEach(() => {
     jest.clearAllMocks()
     pwd.mockReturnValueOnce({ stdout: '/configuration-package' })
@@ -25,6 +17,13 @@ describe('sharec > install', () => {
   })
 
   it('should install configs to the target project', async () => {
+    const targetProcess = {
+      argv: [null, null, 'install'],
+      env: {
+        INIT_CWD: '/target',
+      },
+      exit: jest.fn(),
+    }
     const dir = {
       '/target/.eslintrc': eslintFxt.current,
       '/target/.babelrc': babelFxt.current,
@@ -48,10 +47,54 @@ describe('sharec > install', () => {
 
     expect(vol.readFileSync('/target/.eslintrc', 'utf8')).toWraplessEqual(eslintFxt.result)
     expect(vol.readFileSync('/target/.babelrc', 'utf8')).toWraplessEqual(babelFxt.result)
-    expect(vol.readFileSync('/target/.editorconfig', 'utf8')).toWraplessEqual('bar')
+    expect(vol.readFileSync('/target/.editorconfig', 'utf8')).toWraplessEqual('bar\n')
     expect(vol.readFileSync('/target/.gitignore', 'utf8')).toWraplessEqual(gitignoreFxt.result)
     expect(vol.readFileSync('/target/.npmignore', 'utf8')).toWraplessEqual(npmignoreFxt.result)
     expect(vol.readFileSync('/target/package.json', 'utf8')).toWraplessEqual(packageFxt.result)
     expect(vol.readdirSync('/target/node_modules/.cache/sharec/awesome-config/1.0.0')).toHaveLength(6)
+  })
+
+  it('should place cache in target project if includeCache mode is active', async (done) => {
+    const targetProcess = {
+      argv: [null, null, '--include-cache'],
+      env: {
+        INIT_CWD: '/target',
+      },
+      exit: jest.fn(),
+    }
+    const dir = {
+      '/target/.eslintrc': eslintFxt.current,
+      '/target/.babelrc': babelFxt.current,
+      '/target/.npmignore': npmignoreFxt.current,
+      '/target/.gitignore': gitignoreFxt.current,
+      '/target/package.json': packageFxt.current,
+      '/configuration-package/configs/.eslintrc': eslintFxt.upcoming,
+      '/configuration-package/configs/.editorconfig': 'bar',
+      '/configuration-package/configs/.babelrc': babelFxt.upcoming,
+      '/configuration-package/configs/npmignore': npmignoreFxt.upcoming,
+      '/configuration-package/configs/gitignore': gitignoreFxt.upcoming,
+      '/configuration-package/package.json': JSON.stringify({
+        name: 'awesome-config',
+        version: '1.0.0',
+      }),
+      '/configuration-package/configs/package.json': packageFxt.upcoming,
+    }
+    vol.fromJSON(dir, '/')
+
+    await sharec(targetProcess)
+
+    expect(vol.readFileSync('/target/.eslintrc', 'utf8')).toWraplessEqual(eslintFxt.result)
+    expect(vol.readFileSync('/target/.babelrc', 'utf8')).toWraplessEqual(babelFxt.result)
+    expect(vol.readFileSync('/target/.editorconfig', 'utf8')).toWraplessEqual('bar\n')
+    expect(vol.readFileSync('/target/.gitignore', 'utf8')).toWraplessEqual(gitignoreFxt.result)
+    expect(vol.readFileSync('/target/.npmignore', 'utf8')).toWraplessEqual(npmignoreFxt.result)
+    expect(vol.readFileSync('/target/package.json', 'utf8')).toWraplessEqual(packageFxt.result)
+    expect(vol.readdirSync('/target/.sharec/.cache/awesome-config/1.0.0')).toHaveLength(6)
+
+    try {
+      vol.readdirSync('/target/node_modules/.cache/sharec/awesome-config/1.0.0')
+    } catch (err) {
+      done()
+    }
   })
 })